+++ src/components/map/MapControls.js
@@ -10,7 +10,7 @@ export default function MapControls({ children, className }) {
   return (
     <div className={`c-we-map-controls ${classNames}`}>
       <ul className="map-controls-list">
-        {React.Children.map(children, (ch, i) => (
+        {React.Children.map(children.filter(c => !!c), (ch, i) => (
           <li className="map-controls-item" key={i}>
             {ch}
           </li>